consolidate todo components to separate tsx file
